+++ config-schema.json
@@ -1,0 +1,8 @@
+"json.schemas": [
+    {
+        "fileMatch": [
+            "/config.json"
+        ],
+        "url": "./meta/config.template.json"
+    }
+]
