+++ meta/config.template.json
@@ -3,6 +3,7 @@
     "InternalIPOverride": "1.2.3.4",
     "Starting_Port": 40000,
     "PR_IDLE_WAIT_MS": 15000,
+    "webhookPort": 9193,
     "minPort": 40000,
     "maxPort": 50000,
     "maxConsecutive": 0,
