+++ package.json
@@ -21,5 +21,8 @@
     "download-git-repo": "^3.0.2",
     "express": "^4.17.1",
     "github-api": "^3.4.0"
+  },
+  "devDependencies": {
+    "@types/node": "^14.14.35"
   }
 }
