+++ tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "display": "Node 14",
+
+    "compilerOptions": {
+        "module": "commonjs",
+        "strict": true,
+        "jsx": "preserve",
+        "sourceMap": true,
+        "noImplicitAny": true,
+        "target": "ES2020",
+        "lib": ["ES2020"],
+        "forceConsistentCasingInFileNames": true,
+        "esModuleInterop": true,
+        "skipLibCheck": true,
+        "resolveJsonModule": true
+    },
+    "exclude": [
+        "node_modules",
+        ".vscode",
+        "meta"
+    ],
+    "include": [
+        "src/**/*"
+    ]
+}
